Remove is_valid_password function.
See <https://github.com/agroal/pgagroal/pull/200#issuecomment-1032660243>.
The is_valid_password() was checking only the password length and the fact
that was made by ASCII chars.
The check for the length can be done "inline" directly within a loop.
Added a constant with the minimal length of the password, so that it is possible
to insert a warning message for the user in the case she inputs a too short password.