Bootstrap the generated password.
When automatically generating a password, select a different set of chars
for every single position in the password, so that the password
will have always an upper case, followed by a lower case, followed by
a digit, followed by a special, and so on.

This is not the most secure approach, but generates a quite
complex, and most notably, coherent and valid, master key password.