[#45] Log rotation support.
This is a possible implementation of log rotation.
The idea is to have the following configuration option:
- log_rotation_size expresses the bytes of max size for a file
  before its rotation;
- log_rotation_age expresses the seconds since the last
  rotation.

In order to allow log rotation, log_path must allow for
strftime(3) format strings.
Moreover, there is the need for a set of utility functions
to quickly open and rotate the log file.
The truncation is performed via the already existing
log_mode configuration option.

The parsing of log_roation_size accepts a number (bytes)
and a few suffixes, like 'M' for megabytes.
Similarly, log_rotation_age accepts a number (seconds)
and a few suffixes like 'd' for days.
Only one suffix is supported at a time.

If the logging is done to console or in general not to
a file, the rotation is disabled.

There is a global variable within logginc.g that
keeps track about the current filename as
"expanded" from strftime, so that when
a rotation is required, the file can be flushed
and closed and a new one can be opened.

Close #45