[#44] Support for custom log line prefix.
This introduces the configuration parameter `log_line_prefix`
that accepts a strftime(3) valid string that is parsed
and stored at the very first log entry in the configuration
structure (`config->log_line_prefix`).
Logging to a file and to the console is done using such
prefix via strftime.

Documentation updated.

Close #44