[#224] Prefill a new server when doing a `switch-to`
When a `switch-to` operation is issued, the old server is flushed
(i.e., connections are closed) and new connections to the new primary
server are established.
When such connections are established, the system does a full prefill
against the new primary, so that when the `switch-to` completes, the
new primary is prefilled as the old one was.

Adds also a better detail about the server that is going to be
flushed.

Close #224