[#224] Prefill when doing a `switch-to` or primary change.
Whenever the primary host is changed, by means of an explicit
`switch-to` command or by a primary failure, the connection flushing
is activated. If possible, the prefill should be also restored on the
new server.
As suggested in
<https://github.com/agroal/pgagroal/pull/225#issuecomment-1117218107>
it would be nice to check if the specified new server is the same as
the old one (failure of the primary) or a different one (`switch-to`),
and in the case they are different the prefill is forced.

The prefill will always be to the `INITIAL` size when needed.
If the primary is not set, there is no need to prefill
with the `MIN_SIZE` value.

Close #224