[#223] Keep server state during a `reload`
When doing a `pgagroal-cli reload` the system reloads the
configuration and copies the server information via `copy_server`
function.
This commit makes the `copy_server` to keep the original server status
among a reload process.
This makes a "primary" as "not init (primary)" status, that is surely
better than "not init".