[#230] Correctly read last line of configuration file if no new-line
In the case the main configuration file has no new line in it, the
line is ignored by the function `extract_key_value()`.
This can be a little tricky, since the line is ignored and the user
has no clue about it.
With this fix, the last line of `pgagroal.conf` is always interpreted
even if no new line is present.

Fixes the same problem even for `pgagroal_hba.conf` file read within `extract_hba`.

Clse #230