[#228] No warning messages about pid file when configuration  reloads.
When configuration is reloaded via `pgagroal-cli reload`, the system
checks all the parameters and warns the user about the need of a
restart if some particular parameters have changed.
The check for the `pidfile` parameter is done via `restart_string()`
that, in the case the pidfile has not been specified in the
configuration file, finds a value in the old configuration (the one
running) and no value at all in the new one.
To avoid the misbehaviour, `pgagroal_validate_configuration()` now
invokes a new function `pgagroal_init_pidfile_if_needed()` that sets
the default value of the `pidfile` in the configuration structure.
This way, if the parameter has its default value, no difference will
be spot and no warning message will be issued.

It is safe to make the pidfile default configuration as the last step
in `pgagroal_validate_configuration()` because this function should be
called whenever the configuration is read, and after all this ensures
the configuration is in a "valid" state. On the other hand, it is not
safe to do such configuration within `pgagroal_read_configuration()`
because the default pidfile is built on top of `unix_socket_dir`, that
at that time could be not set.

Close #228