[#232] Keep the same state for the same server during a reload.
This adds a minor refactoring of `transfer_configuration()` and
`copy_server()`.
The memory of the "new" (current) configuration is not anymore
nuked (i.e., zero filled) before `copy_server()` is called, but is
rather done on a per-server slot basis within `copy_server()`.
This allows `copy_server()` to keep track of the previous (i.e.,
running) server state so that it can copy it to the new configuration
server slot.
However, the transfer of the state must not be done always, but only
when the source and destination server looks the same, in particular
have the same connection. After all, we don't want to keep the server
state if the target host or port have changed. To aim this, a specific
function `is_same_server()` has been introduced and used to
"guard" against changes in the connection string.

Adds also a function `restart_server()` that checks for two server
having the same cnnection (by means of `is_same_server()`) and
emits warning messages about a required restart, for example:

   Restart required for Server <venkman>,
   parameter <port> - Existing 5432 New 54322

This also can detect reordering of servers within configuration file.

Partially related to #223
Close #232