[#245] Fix warnings about server sections.
Due to #242, when the `key_in_section()` function checks a paramter
that can be both global and local (e.g., `host`, `port`), the first
call to the function could set the `unknown` flag to true, because the
section does not match, and then the flag remains set even for the
other check.

This commit fixes the problem by not using the `unknown` flag when
dealing with the global section.

Close #245