[#105] Bloecked connections metrics
Creates a new metric named "connection awaiting" to keep track of the
number of connection enqueued but not served due to a reached limit.

This commit introduces a new field in the `prometheus` structure named
`connection_awaiting` to keep track of the connections that have been
suspended due to `blocking_timeout` configuration parameter.
Whenever a connection is suspended by the pool, the
`pgagroal_prometheus_connection_awaiting()` function is invoked to
increment the counter of awaiting connections, and following the
opposite happens with the
`pgagroal_prometheus_connection_unawaiting()` that decrements the
counter, since the connection workflow is going to restart (or abort).

The metric is exported as `pgagroal_connection_awaiting`.

Close #105