Refactor of the prometheus `limit_information()` function.
The prometheus function `limit_information()` prepares the limit
information section within a loop over the limit entries.
I replaced the `append()` calls for every part of the single entry
with a single `snprintf()` with the information for the single entry.

This makes the code shorter and possibly more readble.

Lossely related to #105 and the initial work in #248