[#257] Refactoring of the configuration age and size parsing.
This commit introduces two utility functions:
- `as_seconds()` that parses an "age" string (e.g., "2d") and returns the
   extracted value as seconds;
- `as_bytes()` parses a "size" string (e.g., "2m") and returns the
  number of bytes the string represents.

Such utility functions can be used as backbone for introducing new configuration
parameters, so far they are used as delegates
for `as_log_rotation_age()` and `as_log_rotation_size()`.

Also, converts the `log_rotation_age` and `log_rotation_size` configuration
parameters to `unisgend int`: it does not make any sense to have a negative
rotation value.

In the case a configuration parameter for an age or a size is expressed
as a negative value, a warn about the "unknown" configuration value
is issued.

Documentation updated.

Close #257