[#259] Omit configuration warnings when running cli
`pgagroal-cli` reparses the configuration file as the daemon.
Due to that, if there is a problem with the configuration a warning
is emitted to the stderr.
This commit avoids the `pgagroal-cli` to report warnings on the command line
since such problems could have been alread fixed by the running daemon.

Close #259