[#280] Improve tutorials
This commit makes the tutorials more detailed and ease to follow.

All the tutorials have been expanded to better explain what the aim of
each feature/tutorial is. Moreover, every tutorial points to the other
pieces of the documentation where similar concepts have mentioned or
explained. There is also a more verbose explaination of which user to
use when running commands.
Last, all files have been set to a full path, with the advice that
"standard" file names can be omitted on the command line.

The installation tutorial has been split into two main parts:
- installing and configuring PostgreSQL
- installing and configuring pgagroal.

The PostgreSQL version has been bumped to 14, and the PGDATA has been
set to a not-temporary directory.

Backlink from `GETTING STARTED` to the tutorials.

See #277, #274, #273, #272, #271

Close #280