[#282] Reduce code duplication while reading configuration.
This commit introduces a few constants to indicate the default files
`pgagroal` is going to read the configuration from when the user does
not specify anything else on the command line.

All the constants have the name `PGAGROAL_DEFAULT_<what>_FILE`, like
for example `PGAGROAL_DEFAULT_CONF_FILE` and
`PGAGROAL_DEFAULT_HBA_FILE`.

The idea is to give priority to what the user has specified on the
command line, and then if nothing has been specfiied, use the default
file path. This was already done, but in a more verbose way.

This commit uniforms the error checking for all the cases, when the
file has been specified or it is the default one, so that for
instance, the default file is checked against the same type of
troubles the user could have introduced into custom files.

The online help has been updated to indicate where the default file
for every option is located.

The command line tool `pgagroal-cli` has been adjusted to use the new
constants and get a coherent online help.

Close #282