[#215] Process title length
This is a brute force approach to set the process title with the
appropriate length.
The idea is to compute the size of the process title as it has been
requested, and set it to the computed size.

This works on Rocky Linux release 8.6 (Green Obsidian).

It could break some aggressive security-enforced environments.

This commit introduces the function
`pgagroal_set_connection_proc_title()` that builds a process title
with the form:

      user@host:port/database

using in turn the `pgagroal_set_proc_title` function.
The `pgagroal_set_connection_proc_title` function is used only by the
`worker` process stuff and required to identify the primary server (to
get the hostname and port).

Close #215.