[#296][#297] Improve configuration error management
This commit introduces a check for `pgagroal.conf` to avoid duplicated
sections, missing of main `[pgagroal]` section and the usage of
duplicated server endpoints under different section names.

A new structure, named `struct config_section` has been introduced to
handle the parsing of every section in the configuration file. The aim
of the struct is to temporarily store the line number and the name of
the section as the parsing goes on.
At the end of parsing, all the found sections are compared to see if
there are duplicates or errors.

The function `pgagroal_read_configuration()` has been refactored so
that it returns a negative value in the case of file problem (e.g.,
file not found or too many sections within the file), or a positive
value to indicate how many problematic (e.g., duplicated) sections have been found.

The `main` function has been refactored to adapt to the new return
values and to provide better descriptive messages to the user. In
particular a new temporary variable `message` is now used to format a
coherent error message between the application output and the systemd
notification.

Moreover, when validating the main configuration, there is a nested
loop to check against duplicated servers, i.e., servers that have the
very same endpoint (host and port). The check relies on
`is_same_server' and therefore it has all the limitation that the
latter has, most notably it does not performs a DNS lookup, so at the
moment the usage of same servers with different hostnames cannot be
detected.
The detection of duplicated server is especially useful to allow
`switch-to` and the failover to not perform useless operations.

Several constants have been defined:
- PGAGROAL_CONFIGURATION_STATUS_FILE_TOO_BIG if the file contains
too much sections or users or entries, (e.g., for sections if it
overflows  NUMBER_OF_SERVERS + 1);
- PGAGROAL_CONFIGURATION_STATUS_FILE_NOT_FOUND to indicate that the
file does not exists;
- PGAGROAL_CONFIGURATION_STATUS_CANNOT_DECRYPT to indicate a mastre
key problem;
- PGAGROAL_CONFIGURATION_STATUS_KO to indicate a generic error;
- PGAGROAL_CONFIGURATION_STATUS_OK to indicated everything is fine.

Since now there are constants to define the return values for
different error conditions in the configuration parsing, the `main`
and the `pgagroal-cli` have been refactored to use such return values
in a more readable and useful way.

Documentation of the functions have been updated.
User documentation updated to better explain limits on the number of
sections. Also, place brief documentation to explain every
configuration file limits.

All error paths in configuration read functions have been merged into
a single one. Since calling `free(3)` on a NULL pointer
has no effect at all, it is safe to merge all the error paths
in a single one and keep track only of the return value to emit.

This improves code readability.

Close #296.
Close #297