[#298][#299] Improve management of pid file
In `main` the pidfile is created, then it could happen that the
application brutally exits because of a runtime error, like not being
able to open a socket. In such cases, the pid file has to be removed.

This commit introduces an `error` path that does remove the pidfile
and exits with an error code. Invoking `remove_pidfile` is safe
even when the file does not exists, and besides the pid file cannot be
changed at runtime via a configuration reload.

Moreover, it is not worth checking if the pidfile exists on disk before
deleting it via `unlink(2)`, since the function will run smoothly if
the file is not there.

[#298] Remove pidfile when reloading configuration fails.

Change the permissions on the pid file: 0640.

Since the configuration reload is issued within the main process, the
process must remove the pidfile (if any) before existing in case of
failure.

This commit also removes `access(2)` call before `open(2)` in pidfile creation

Since `open(2)` is invoked with flags `O_CREAT | O_EXCL`, the function
will fail if the file is already there. This means there is no need
for `access(2)` to be invoked before the `open(2)` call.

This commit catches errors related to `open` not being able to create
the pid file because another one is there, or because the user does
not have permissions.
Other errors are managed in a generic way as it was before.

Close #298
Close #299