[#319] Remove `atoi` from `pgagroal-cli`
When doing a remote connection to the pooler, the user has to specify
the port with `-p` flag. The value is converted with `strtol` and the
ranges are checked.
Also, since pgagroal cannot be run as root, it cannot bind to ports
lower than 1024, therefore if the user specifies a wrong port number,
it gets a different error message rather than "No route to host".

Close #319