[#330] Fix exit code handling in `pgagroal_cli`
This commit fixes some minor socket-driven communications where the
return code was not propagated correctly to the application exit code,
and therefore some erros could get a wrong exit code marked as
"success".

Meanwhile, little changes the error messages for a failure exit code,
since a failure does not automatically imply a "no connection" status,
rather could be a "connection problem" (i.e., some data has been
sent/received over the socket, some did not).

Last, reduce the code to print the status of the command when using
the `--verbose` flag. Use `warnx(3)` to print the verbose exit
message, so that it can be redirected to another place since it is
printed on STDERR.

Close #330