[#335] Disambiguate connection errors from data errors in commands
This commit introduces a few constants to indicate if the
communication over the management socket has failed due to a
connection error or because the data received is not valid.
So far, only `config-get` returns a different value to indicate
invalid data over the socket, but other commands can take advanatge of
this too.

The `pgagroal-cli` main loop has been refactored to distinguish the
new cases, and now it prints the "connection error" message only when
the low level function reports such an error.

All the functions now return a specific value among those defined.

Close #335