[#337] Fix compilation problems
As reported in CI test failures
<https://github.com/agroal/pgagroal/pull/336#issuecomment-1348346821>
and
<https://github.com/agroal/pgagroal/pull/336#issuecomment-1349502055>
the system does not compile under GCC 12 with `-Werror` option.

This commit fixes the string length problem reported by gcc.
The `to_string` internal method used to get a stringify value
of a configuration parameter has changed so that now accepts a buffer
size. Such size indicates the max length of the destinatiojn buffer
and allows the function to do some more checks about what to copy onto
the buffer.

Close #337