[#339] Count how many settings remain unchanged in configuration reload.
The internal function `transfer_configuration` now returns a negative
value in the case of unchanged settings: such value counts the number
of parameters that has remained untouched, for example because they
require a restart.
The idea is that there are parameter that cannot be changed on the fly
and that do not prevent the configuration to be swapped, but it could
be interesting to take the count of unchanged configuration settings.

Close #339