[#355] Add line number field in various configuration structs
Added a `lineno` field into structures `struct hba` and `struct
server`.
This simplifies error reporting, since it is now possible to drive the
user to the line number that is causing a problem.

Thanks to this, it is possible to provide better messages to the users
about misconfigurations.
For example, if the user mispells an entry into the HBA file:

```
FATAL configuration.c:1162 Unknown HBA type: HOSTSS (/etc/pgagroal/pgagroal_hba.conf:7)
```

Changes to a few log/warn messages to become more easy to read.

Close #355